+++ owncloudApp/src/main/java/com/owncloud/android/providers/TransferProvider.kt
@@ -2,36 +2,38 @@
  * ownCloud Android client application
  *
  * @author Abel García de Prada
- *
  * Copyright (C) 2020 ownCloud GmbH.
- *
+ * <p>
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 2,
  * as published by the Free Software Foundation.
- *
+ * <p>
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
- *
+ * <p>
  * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http:></http:>//www.gnu.org/licenses/>.
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
 package com.owncloud.android.providers
 
-import android.content.Context
-import com.owncloud.android.utils.ConnectivityUtils
+import android.accounts.Account
+import androidx.work.WorkManager
+import com.owncloud.android.domain.files.model.OCFile
+import java.util.UUID
 
-class OCContextProvider(private val context: Context) : ContextProvider {
+/**
+ * Facade to execute transfer operations. Should be only handled by Transfer Manager.
+ */
+interface TransferProvider {
 
-    override fun getBoolean(id: Int): Boolean = context.resources.getBoolean(id)
-
-    override fun getString(id: Int): String = context.resources.getString(id)
-
-    override fun getContext(): Context = context
-
-    override fun isConnected(): Boolean {
-        return ConnectivityUtils.isAppConnected(context)
-    }
+    /**
+     * Enqueue a new download and return its uuid.
+     * You can check and observe its progress using
+     * @see WorkManager.getWorkInfoById and
+     * @see WorkManager.getWorkInfoByIdLiveData
+     */
+    fun downloadFile(account: Account, file: OCFile): UUID?
 }
